requirements.txt
evernote
pyyaml
click
